create separate soap auth headers for curam atp and mec check
