+++ spec/domain/operations/curam/encode_account_transfer_check_request_spec.rb
@@ -15,7 +15,7 @@ describe Curam::EncodeAccountTransferCheckRequest, "given a Curam::AccountTransf
   let(:operation) { Curam::EncodeAccountTransferCheckRequest.new }
   let(:header) do
     instance_double(
-      Curam::SoapAuthorizationHeader,
+      Curam::Atp::SoapAuthorizationHeader,
       {
         :username => "A Username",
         :password => "A Password",
